requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
sqlalchemy==2.0.23
qrcode[pil]==7.4.2
Pillow==10.1.0
pydantic==2.5.0
pydantic-core==2.14.1
python-multipart==0.0.6
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0
email-validator==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1